Generalize application of visuals to CSS
